// src/types.ts
export interface DanfsePessoa {
    /** CNPJ formatado ou CPF formatado ou NIF (texto livre) */
    doc: string;
    /** Inscrição Municipal */
    IM: string;
    /** Telefone */
    fone: string;
    /** Nome / Nome Empresarial */
    xNome: string;
    /** Município / Sigla UF (ex: "Formiga / MG") */
    municipioUf: string;
    /** Código IBGE / CEP (ex: "3126109 / 35.570-000") */
    ibgeCep: string;
    /** Endereço concatenado (xLgr, nro, xCpl, xBairro) */
    endereco: string;
    /** E-mail */
    email: string;
}